+++ tablespecs/xbt_peers_history.sql
@@ -22,7 +22,7 @@
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE IF NOT EXISTS `xbt_peers_history` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(24) NOT NULL AUTO_INCREMENT,
   `uid` int(11) NOT NULL,
   `downloaded` bigint(20) NOT NULL,
   `remaining` bigint(20) NOT NULL,
